add : cpp_programming_questions 10

+++ 01_basics_of_c++/04_area_of_circle.cpp
@@ -10,7 +10,7 @@ using namespace std;
 // // Main Function Start
 int main()
 {
-    float radius, area;
+    double radius, area;
 
     cout << "\nEnter Radius of Circle => ";
     cin >> radius;
